pyproject.toml
[tool.poetry]
name = "cv-service"
version = "0.1.0"
description = "Computer Vision Service for Classification and Detection"
authors = ["Your Team <team@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
python-multipart = "^0.0.6"
pillow = "^10.0.0"
numpy = "^1.24.0"
torch = "^2.0.0"
torchvision = "^0.15.0"
ultralytics = "^8.0.0"
hydra-core = "^1.3.0"
omegaconf = "^2.3.0"
python-dotenv = "^1.0.0"
opencv-python-headless = "^4.8.0"
loguru = "^0.7.0"
redis = "^4.6.0"
fastapi-limiter = "^0.1.5"
allenai-common = "1.1.2"
pytest = "7.4.0"
pytest-asyncio = "0.21.1"
pytest-cov = "4.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.4.1"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true 